Adjust translations so sprites render. Clear to black.
Through trial and error I discovered two adjustments that would cause
paddles to start rendering.  1) The camera needed to be shifted in the
opposite x and y directions. 2) The paddle sprites needed to have
negative z values.  This second problem was especially irritating, as I
can't figure out _why_ it needs to be done. I just stumbled onto it.

I set the screen to black as well. The rest of the changes are just
rustfmt.